config/feat: move remark plugins into folder
