+++ astro.config.mjs
@@ -8,8 +8,8 @@ import vercel from "@astrojs/vercel/serverless"
 import compress from "astro-compress"
 import robots from "astro-robots-txt"
 
-import { remarkReadingTime } from "./remark-reading-time.mjs"
-import { remarkWidont } from "./remark-widont.mjs"
+import { remarkReadingTime } from "./remark-plugins/remark-reading-time.mjs"
+import { remarkWidont } from "./remark-plugins/remark-widont.mjs"
 
 // https://astro.build/config
 export default defineConfig({
